styling changes, begin work on login
