+++ rpgains/src/app/app.component.css
@@ -1,43 +1,2 @@
-#navbar {
-  overflow: hidden;
-  /*background-color: #46344E;*/
-  position: fixed;
-  top: 0;
-  width: 100%;
-  z-index: 1;
-  left: 0;
-  right: 0;
-}
 
 
-#navbar img {
-  /*display: block;*/
-  margin-left: auto;
-  margin-right: auto;
-  color: #faed26;
-  text-align: center;
-  display: inline-block;
-  padding: 10px 10px 10px 10px;
-  vertical-align: middle;
-}
-
-#navbar a {
-  float: right;
-  display: block;
-  color: #faed26;
-  text-align: center;
-  padding: 25px 25px 0 0;
-  text-decoration: none;
-  font-size: 25px;
-}
-
-#navbar a:hover {
-  background: #5a5560;
-  color: #fff;
-}
-
-
-
-body{
-  padding-top: 65px;
-}
