+++ rpgains/src/app/app.component.html
@@ -1,8 +1,3 @@
-<nav id="navbar">
-  <img src="../assets/rpgLogo2.png" height="75" routerLink="/homepage"/>
-  <a routerLink="/login">Sign-in/Up</a>
-  <a routerLink="/dashboard">Play</a>
-</nav>
 <router-outlet></router-outlet>
