+++ rpgains/src/app/homepage/homepage.component.html
@@ -1,3 +1,15 @@
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+<nav id="navbar">
+  <img src="../assets/rpgLogo2.png" height="75" routerLink="/homepage"/>
+  <a routerLink="/login" id="logged">Sign-in/Up</a>
+
+  <a routerLink="/dashboard">Play</a>
+</nav>
+<script>
+
+</script>
+
 <div class="bgimg-1">
   <div class="caption">
 
@@ -15,9 +27,12 @@
   <div class="section">
     <p style="font-size: 20px">RPGains was created for one reason. To help you get fit.<br>
       We know that motivation can be one of the hardest parts of going to the gym and repeatedly going back.
-      What we've set out to do is create another reason for you to go to the gym, track your fitness, and reach your goals.
-      RPGains is a game-based aide to your workout. It provides a simple way to track your workout, progress, and goals, and rewards you for your persistence.
-      By giving incentives for going to gym, as well as meeting your calorie goals, and increasing your strength training, users may strive towards a healthier
+      What we've set out to do is create another reason for you to go to the gym, track your fitness, and reach your
+      goals.
+      RPGains is a game-based aide to your workout. It provides a simple way to track your workout, progress, and goals,
+      and rewards you for your persistence.
+      By giving incentives for going to gym, as well as meeting your calorie goals, and increasing your strength
+      training, users may strive towards a healthier
       lifestyle, and a stronger body.
     </p>
   </div>
