+++ rpgains/src/app/login/login.component.html
@@ -1,14 +1,90 @@
+<html>
+<meta charset="utf-8">
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<head>
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+  <script src="../../services/mongoService.js"></script>
+</head>
+<script src="../../services/mongoService.js"></script>
+<body>
+<nav id="navbar">
+  <img src="../assets/rpgLogo2.png" height="75" routerLink="/homepage"/>
+  <a routerLink="/login" id="logged">Sign-in/Up</a>
+  <a routerLink="/dashboard">Play</a>
+</nav>
+<script>
+  if (localStorage.session == 'logged') {
+    document.getElementById('logged').innerHTML = "<button>Log out</button>"
+  }
+</script>
+<!-- NEW USER-->
 <section class="page-section cta">
-  <div class="container">
+  <div class="container-fluid">
     <div class="row">
-      <div class="col-xl-6 mx-auto">
+      <div class="col-sm-5 mx-auto">
         <div class="cta-inner text-center rounded">
           <h2 class="section-heading mb-5">
             <span class="section-heading-upper">Join Us!</span>
             <span class="section-heading-lower">New Users</span>
           </h2>
           <div class="form-style-5">
-            <form id="newUser" onsubmit="createUser()" class="test1">
+            <form id="newUser" ng-click="testConnection()" class="test1">
+              <input type="text" name="name" placeholder="Your full name" required>
+              <input type="text" name="phone" placeholder="Phone 1+(xxx)xxx-xxxx" required>
+              <script type="text/javascript">
+                document.getElementById('newUser').phone.addEventListener('input', function (e) {
+                  var x = e.target.value.replace(/\D/g, '').match(/(\d{0,3})(\d{0,3})(\d{0,4})/);
+                  e.target.value = !x[2] ? x[1] : '(' + x[1] + ') ' + x[2] + (x[3] ? '-' + x[3] : '');
+                });
+              </script>
+              <input type="email" name="email" placeholder="E-mail" required>
+              <input type="password" name="password" placeholder="Password" required>
+              <input type="password" name="passrepeat" placeholder="Re-type Password" required>
+              <input type="checkbox" name="terms" value="terms" required>Do you accept our terms and conditions?<br><br>
+              <button ng-click="login()" class="btn-primary" type="submit" name="submitButton"></button>
+              <input type="submit" name="submitButton" class="btn-primary" >
+            </form>
+          </div>
+          <br>
+          <small>Issues? support@rpgains.com</small>
+        </div>
+      </div>
+
+      <!-- EXISTING USER-->
+
+      <div class="col-sm-5 mx-auto">
+        <div class="cta-inner text-center rounded">
+          <h2 class="section-heading mb-5">
+            <span class="section-heading-upper">Welcome Back!</span>
+            <span class="section-heading-lower">Existing Users</span>
+          </h2>
+          <div class="form-style-5">
+            <form id="loginUser" onsubmit="login()" class="test1">
+              <input type="email" name="email" placeholder="E-mail" required>
+              <input type="password" name="password" placeholder="Password" required>
+              <input type="submit" name="submitButton" class="btn-primary">
+            </form>
+          </div>
+          <br>
+          <small>Issues? support@rpgains.com</small>
+        </div>
+      </div>
+    </div>
+  </div>
+</section>
+<!--<section class="page-section cta">
+  <div class="container">
+    <div class="row">
+      <div class="col-sm">
+        <div class="cta-inner text-center rounded">
+          <h2 class="section-heading mb-5">
+            <span class="section-heading-upper">Join Us!</span>
+            <span class="section-heading-lower">New Users</span>
+          </h2>
+          <div class="form-style-5">
+            <form id="newUser" onsubmit="testConnection()" class="test1">
               <input type="text" name="name" placeholder="Your full name" required>
               <input type="text" name="phone" placeholder="Phone 1+(xxx)xxx-xxxx" required>
               <script type="text/javascript">
@@ -28,7 +104,8 @@
           <small>Issues? support@rpgains.com</small>
         </div>
       </div>
-      <div class="col-xl-6 mx-auto">
+      <br>
+      <div class="col-sm">
         <div class="cta-inner text-center rounded">
           <h2 class="section-heading mb-5">
             <span class="section-heading-upper">Welcome Back!</span>
@@ -50,4 +127,6 @@
       </div>
     </div>
   </div>
-</section>
+</section>-->
+</body>
+</html>
