+++ rpgains/src/app/login/login.component.ts
@@ -9,6 +9,12 @@ export class LoginComponent implements OnInit {
 
   constructor() {}
 
+  logout() : void{
+    localStorage.removeItem('session');
+    localStorage.removeItem('loggedUser');
+    localStorage.removeItem('userName');
+  }
+
   ngOnInit() {
   }
