+++ rpgains/src/styles.css
@@ -1,7 +1,69 @@
 /* You can add global styles to this file, and also import other style files */
 html{
   background-color: #46344E;
+
 }
 body{
   padding-top: 75px;
+  font-family: 'Oswald', sans-serif;
 }
+
+#navbar {
+  overflow: hidden;
+  background-color: #46344E;
+  position: fixed;
+  top: 0;
+  width: 100%;
+  z-index: 1;
+  left: 0;
+  right: 0;
+}
+
+
+#navbar img {
+  /*display: block;*/
+  margin-left: auto;
+  margin-right: auto;
+  color: #faed26;
+  text-align: center;
+  display: inline-block;
+  padding: 10px 10px 10px 10px;
+  vertical-align: middle;
+}
+
+#navbar a {
+  float: right;
+  display: block;
+  color: #ffffff;
+  text-align: center;
+  padding: 25px 25px 0 0;
+  text-decoration: none;
+  font-size: 25px;
+  text-shadow:
+    -1px -1px 0 #000000,
+    1px -1px 0 #000000,
+    -1px 1px 0 #000000,
+    1px 1px 0 #000000;
+}
+
+#navbar div{
+  float: right;
+  display: block;
+  color: #ffffff;
+  text-align: center;
+  padding: 25px 25px 0 0;
+  text-decoration: none;
+  font-size: 25px;
+  text-shadow:
+    -1px -1px 0 #000000,
+    1px -1px 0 #000000,
+    -1px 1px 0 #000000,
+    1px 1px 0 #000000;
+}
+
+
+#navbar a:hover {
+  background: #5a5560;
+  color: #fff;
+}
+
